refactor(host-domain): split monolithic lib.rs into focused domain modules
Extract task, document, system, git, runtime, and store modules.

Keep public API stable via lib.rs re-exports and retain existing tests.
