+++ apps/desktop/src-tauri/crates/host-domain/src/git.rs
@@ -1,0 +1,60 @@
+use anyhow::Result;
+use serde::{Deserialize, Serialize};
+use std::path::Path;
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct GitBranch {
+    pub name: String,
+    pub is_current: bool,
+    pub is_remote: bool,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct GitCurrentBranch {
+    pub name: Option<String>,
+    pub detached: bool,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct GitWorktreeSummary {
+    pub branch: String,
+    pub worktree_path: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct GitPushSummary {
+    pub remote: String,
+    pub branch: String,
+    pub output: String,
+}
+
+pub trait GitPort: Send + Sync {
+    fn get_branches(&self, repo_path: &Path) -> Result<Vec<GitBranch>>;
+    fn get_current_branch(&self, repo_path: &Path) -> Result<GitCurrentBranch>;
+    fn switch_branch(
+        &self,
+        repo_path: &Path,
+        branch: &str,
+        create: bool,
+    ) -> Result<GitCurrentBranch>;
+    fn create_worktree(
+        &self,
+        repo_path: &Path,
+        worktree_path: &Path,
+        branch: &str,
+        create_branch: bool,
+    ) -> Result<()>;
+    fn remove_worktree(&self, repo_path: &Path, worktree_path: &Path, force: bool) -> Result<()>;
+    fn push_branch(
+        &self,
+        repo_path: &Path,
+        remote: &str,
+        branch: &str,
+        set_upstream: bool,
+        force_with_lease: bool,
+    ) -> Result<GitPushSummary>;
+}
