+++ apps/desktop/src-tauri/crates/host-domain/src/runtime.rs
@@ -1,0 +1,92 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum RunState {
+    Starting,
+    Running,
+    Blocked,
+    AwaitingDoneConfirmation,
+    Completed,
+    Failed,
+    Stopped,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RunSummary {
+    pub run_id: String,
+    pub repo_path: String,
+    pub task_id: String,
+    pub branch: String,
+    pub worktree_path: String,
+    pub port: u16,
+    pub state: RunState,
+    pub last_message: Option<String>,
+    pub started_at: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct AgentRuntimeSummary {
+    pub runtime_id: String,
+    pub repo_path: String,
+    pub task_id: String,
+    pub role: String,
+    pub working_directory: String,
+    pub port: u16,
+    pub started_at: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(tag = "type")]
+#[serde(rename_all = "snake_case")]
+pub enum RunEvent {
+    RunStarted {
+        run_id: String,
+        message: String,
+        timestamp: String,
+    },
+    AgentThought {
+        run_id: String,
+        message: String,
+        timestamp: String,
+    },
+    ToolExecution {
+        run_id: String,
+        message: String,
+        timestamp: String,
+    },
+    PermissionRequired {
+        run_id: String,
+        message: String,
+        command: Option<String>,
+        timestamp: String,
+    },
+    PostHookStarted {
+        run_id: String,
+        message: String,
+        timestamp: String,
+    },
+    PostHookFailed {
+        run_id: String,
+        message: String,
+        timestamp: String,
+    },
+    ReadyForManualDoneConfirmation {
+        run_id: String,
+        message: String,
+        timestamp: String,
+    },
+    RunFinished {
+        run_id: String,
+        message: String,
+        timestamp: String,
+        success: bool,
+    },
+    Error {
+        run_id: String,
+        message: String,
+        timestamp: String,
+    },
+}
